refactor(maps): extract useMapDropTarget hook from ActiveMapStage
Move the wiki drag-and-drop → create-token logic (incl. the shift-drag batch
placement and the drag-over highlight state) into a self-contained
useMapDropTarget hook. No behavior change; verified by the monster-tokens e2e.
